adding pet profile pictures to the pets tab for post, and adding them to the get requests on the booking step modal.
adding user profile get requests to the conversations and messages screens.

adding pet id to the response of the add-pet.

adding backend support for the base64 photos for pets on the backend.

removing active button from the pro service cards.

adding to lowercase and first letter upper case to the pets names and species and types.

fixing the zoom on profilephotocropper to be smooth slider